Split contexts
