+++ src/sentry/static/sentry/app/components/events/interfaces/contexts/default.jsx
@@ -1,0 +1,18 @@
+import React from 'react';
+
+import ContextBlock from './contextBlock';
+
+const DefaultContextType = React.createClass({
+  propTypes: {
+    alias: React.PropTypes.string.isRequired,
+    data: React.PropTypes.object.isRequired,
+  },
+
+  render() {
+    return (
+      <ContextBlock data={this.props.data} alias={this.props.alias} />
+    );
+  }
+});
+
+export default DefaultContextType;
